Move confirm publish dialog to file and test in isolation
