+++ packages/core/admin/admin/src/content-manager/pages/ListView/components/BulkActionButtons/ConfirmBulkActionDialog/index.js
@@ -1,9 +1,21 @@
 import * as React from 'react';
 
-import { Button, Flex, Dialog, DialogBody, DialogFooter } from '@strapi/design-system';
-import { ExclamationMarkCircle } from '@strapi/icons';
+import { Button, Flex, Dialog, DialogBody, DialogFooter, Typography } from '@strapi/design-system';
+import {
+  useTableContext,
+  useFetchClient,
+  useNotification,
+  useAPIErrorHandler,
+} from '@strapi/helper-plugin';
+import { Check, ExclamationMarkCircle } from '@strapi/icons';
 import PropTypes from 'prop-types';
 import { useIntl } from 'react-intl';
+import { useQuery } from 'react-query';
+import { useSelector } from 'react-redux';
+
+import InjectionZoneList from '../../../../../components/InjectionZoneList';
+import { getTrad } from '../../../../../utils';
+import { listViewDomain } from '../../../selectors';
 
 const ConfirmBulkActionDialog = ({ onToggleDialog, isOpen, dialogBody, endAction }) => {
   const { formatMessage } = useIntl();
@@ -51,4 +63,102 @@ export const confirmDialogsPropTypes = {
   onToggleDialog: PropTypes.func.isRequired,
 };
 
-export default ConfirmBulkActionDialog;
+/* -------------------------------------------------------------------------------------------------
+ * BoldChunk
+ * -----------------------------------------------------------------------------------------------*/
+
+const BoldChunk = (chunks) => <Typography fontWeight="bold">{chunks}</Typography>;
+
+/* -------------------------------------------------------------------------------------------------
+ * ConfirmDialogPublishAll
+ * -----------------------------------------------------------------------------------------------*/
+
+const ConfirmDialogPublishAll = ({ isOpen, onToggleDialog, isConfirmButtonLoading, onConfirm }) => {
+  const { formatMessage } = useIntl();
+  const { get } = useFetchClient();
+  const { selectedEntries } = useTableContext();
+  const toggleNotification = useNotification();
+  const { formatAPIError } = useAPIErrorHandler(getTrad);
+  const {
+    contentType: { uid: slug },
+  } = useSelector(listViewDomain());
+
+  const {
+    data: countDraftRelations,
+    isLoading,
+    isError,
+  } = useQuery(
+    ['content-manager', 'draft-relations', slug, selectedEntries],
+    async () => {
+      const {
+        data: { data },
+      } = await get(
+        `/content-manager/collection-types/${slug}/actions/countManyEntriesDraftRelations`,
+        {
+          params: {
+            ids: selectedEntries,
+          },
+        }
+      );
+
+      return data;
+    },
+    {
+      onError(error) {
+        toggleNotification({ type: 'warning', message: formatAPIError(error) });
+      },
+    }
+  );
+
+  if (isError) {
+    return null;
+  }
+
+  return (
+    <ConfirmBulkActionDialog
+      isOpen={isOpen && !isLoading}
+      onToggleDialog={onToggleDialog}
+      dialogBody={
+        <>
+          <Typography id="confirm-description" textAlign="center">
+            {countDraftRelations > 0 &&
+              formatMessage(
+                {
+                  id: getTrad(`popUpwarning.warning.bulk-has-draft-relations.message`),
+                  defaultMessage:
+                    '<b>{count} {count, plural, one { relation } other { relations } } out of {entities} { entities, plural, one { entry } other { entries } } {count, plural, one { is } other { are } }</b> not published yet and might lead to unexpected behavior. ',
+                },
+                {
+                  b: BoldChunk,
+                  count: countDraftRelations,
+                  entities: selectedEntries.length,
+                }
+              )}
+            {formatMessage({
+              id: getTrad('popUpWarning.bodyMessage.contentType.publish.all'),
+              defaultMessage: 'Are you sure you want to publish these entries?',
+            })}
+          </Typography>
+          <InjectionZoneList area="contentManager.listView.publishModalAdditionalInfos" />
+        </>
+      }
+      endAction={
+        <Button
+          onClick={onConfirm}
+          variant="secondary"
+          startIcon={<Check />}
+          loading={isConfirmButtonLoading}
+        >
+          {formatMessage({
+            id: 'app.utils.publish',
+            defaultMessage: 'Publish',
+          })}
+        </Button>
+      }
+    />
+  );
+};
+
+ConfirmDialogPublishAll.propTypes = confirmDialogsPropTypes;
+
+export { ConfirmDialogPublishAll, ConfirmBulkActionDialog };
