+++ packages/core/admin/admin/src/content-manager/pages/ListView/components/BulkActionButtons/tests/index.test.js
@@ -4,29 +4,13 @@ import { lightTheme, ThemeProvider } from '@strapi/design-system';
 import { Table, useTableContext } from '@strapi/helper-plugin';
 import { render, screen, waitFor, within } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import { rest } from 'msw';
-import { setupServer } from 'msw/node';
 import { IntlProvider } from 'react-intl';
-import { QueryClient, QueryClientProvider } from 'react-query';
 import { Provider } from 'react-redux';
 import { MemoryRouter } from 'react-router-dom';
 import { combineReducers, createStore } from 'redux';
 
+import BulkActionButtons from '..';
 import reducers from '../../../../../../reducers';
-import BulkActionButtons, { ConfirmDialogPublishAll } from '../SelectedEntriesModal';
-
-const handlers = [
-  rest.get('*/countManyEntriesDraftRelations', (req, res, ctx) => {
-    return res.once(
-      ctx.status(200),
-      ctx.json({
-        data: 0,
-      })
-    );
-  }),
-];
-
-const server = setupServer(...handlers);
 
 const toggleNotification = jest.fn();
 
@@ -68,33 +52,23 @@ const store = createStore(rootReducer, {
 const setup = (props) => ({
   ...render(<BulkActionButtons {...props} />, {
     wrapper({ children }) {
-      const client = new QueryClient({
-        defaultOptions: {
-          queries: {
-            retry: false,
-          },
-        },
-      });
-
       return (
-        <QueryClientProvider client={client}>
-          <ThemeProvider theme={lightTheme}>
-            <IntlProvider locale="en" messages={{}} defaultLocale="en">
-              <Provider store={store}>
-                <MemoryRouter>
-                  <Table.Root>{children}</Table.Root>
-                </MemoryRouter>
-              </Provider>
-            </IntlProvider>
-          </ThemeProvider>
-        </QueryClientProvider>
+        <ThemeProvider theme={lightTheme}>
+          <IntlProvider locale="en" messages={{}} defaultLocale="en">
+            <Provider store={store}>
+              <MemoryRouter>
+                <Table.Root>{children}</Table.Root>
+              </MemoryRouter>
+            </Provider>
+          </IntlProvider>
+        </ThemeProvider>
       );
     },
   }),
 });
 
 describe('BulkActionsBar', () => {
-  it('should render publish buttons if showPublish is true', () => {
+  it('should render publish buttons if showPublish is true', async () => {
     setup({ showPublish: true });
 
     expect(screen.getByRole('button', { name: /\bPublish\b/ })).toBeInTheDocument();
@@ -184,122 +158,5 @@ describe('BulkActionsBar', () => {
     );
 
     expect(onConfirmUnpublishAll).toHaveBeenCalledWith([1, 2]);
-  });
-});
-
-const setupConfirmPublish = () => ({
-  ...render(
-    <ConfirmDialogPublishAll
-      isOpen
-      onConfirm={jest.fn()}
-      onToggleDialog={jest.fn()}
-      isConfirmButtonLoading
-    />,
-    {
-      wrapper({ children }) {
-        const client = new QueryClient({
-          defaultOptions: {
-            queries: {
-              retry: false,
-            },
-          },
-        });
-
-        return (
-          <QueryClientProvider client={client}>
-            <ThemeProvider theme={lightTheme}>
-              <IntlProvider locale="en" messages={{}} defaultLocale="en">
-                <Provider store={store}>
-                  <MemoryRouter>
-                    <Table.Root>{children}</Table.Root>
-                  </MemoryRouter>
-                </Provider>
-              </IntlProvider>
-            </ThemeProvider>
-          </QueryClientProvider>
-        );
-      },
-    }
-  ),
-});
-
-describe('ConfirmDialogPublishAll', () => {
-  beforeAll(() => {
-    server.listen();
-  });
-
-  afterAll(() => {
-    server.close();
-  });
-
-  it('should show a default message if there are not draft relations', async () => {
-    setupConfirmPublish();
-    await waitFor(() => expect(screen.getByRole('dialog')).toBeInTheDocument());
-    expect(
-      screen.queryByText('not published yet and might lead to unexpected behavior')
-    ).not.toBeInTheDocument();
-
-    expect(
-      await screen.findByText('Are you sure you want to publish these entries?')
-    ).toBeInTheDocument();
-  });
-
-  it('should show the warning message with just 1 draft relation and two entries', async () => {
-    server.use(
-      rest.get('*/countManyEntriesDraftRelations', (req, res, ctx) => {
-        return res.once(
-          ctx.status(200),
-          ctx.json({
-            data: 1,
-          })
-        );
-      })
-    );
-
-    setupConfirmPublish();
-    await waitFor(() => expect(screen.getByRole('dialog')).toBeInTheDocument());
-    const draftWarning = await screen.getByText(/relation out of/i);
-    expect(draftWarning.textContent).toBe('1  relation  out of 2  entries   is ');
-  });
-
-  it('should show the warning message with 2 draft relations and two entries', async () => {
-    server.use(
-      rest.get('*/countManyEntriesDraftRelations', (req, res, ctx) => {
-        return res.once(
-          ctx.status(200),
-          ctx.json({
-            data: 2,
-          })
-        );
-      })
-    );
-
-    setupConfirmPublish();
-    await waitFor(() => expect(screen.getByRole('dialog')).toBeInTheDocument());
-    const draftWarning = await screen.getByText(/relations out of/i);
-    expect(draftWarning.textContent).toBe('2  relations  out of 2  entries   are ');
-  });
-
-  it('should not show the Confirmation component if there is an error coming from the API', async () => {
-    server.use(
-      rest.get('*/countManyEntriesDraftRelations', (req, res, ctx) => {
-        return res.once(
-          ctx.status(500),
-          ctx.json({
-            errorMessage: 'Error',
-          })
-        );
-      })
-    );
-
-    setupConfirmPublish();
-    await waitFor(() => expect(screen.queryByRole('dialog')).not.toBeInTheDocument());
-    await waitFor(() =>
-      expect(toggleNotification).toHaveBeenCalledWith({
-        type: 'warning',
-        message: 'Request failed with status code 500',
-      })
-    );
-    expect(await screen.getByRole('alert')).toBeInTheDocument();
   });
 });
